use full nanosecond resolution, segment by thread
If two threads generate a UUID within in the same nanosecond,
the order of creation is arbitrary.
Rather than lock on a mutex by default, segment by thread_local
which trades an increase in through-put for a slight decrease in
order certainty.
Using a mutex is still an option, if desired.

Also add robustness for clocks going backwards in time.